// cypress/PageObject/enums/enums.ts
export enum Timeout{
    SHORT = 2000,
    MEDIUM = 5000,
    LONG = 10000
}
export enum FlightLocation {
      FROM = 'from',
      TO = 'to'
}
export enum AirportName {
    DEPARTURE = "Dublin",
    DESTINATION = "London Stansted",
} 
export enum CalculatedDays{
    // 62 133
    DepartureDays = 69,
    ArrivalDays = 139   

}